Problem 20 Solution
This solution works so long as the input decimal is not an exponent of 16 greater than 1 (for example, 256 4096, etc). I am thinking of creating another solution that works by identifying the highest power term first, then the 2nd highest power term and so on until all terms are identified, after which they may be simply printed one after the other. This should resolve my current problem.