Convert a Binary Number to a Decimal Number
This solution seems to work for all binary numbers that are within the acceptable range of the Scanner function. The solution works by converting the input int into a string so that each digit may be more easily manipulated (technically chars given that they have been converted to a string). The degree of the binary number is identified, which allows the mapping of each digit of the binary number to an exponent. After that, we simply loop through each digit of the binary number to see if it is a 1, in which case we simply add 2^itsDegree to another decimal number variable to get the final result.