Convert a binary number to hexadecimal number
This solution was made be reusing code from my solutions to problem 20 (convert a decimal number to hexadecimal) and 22 (convert a binary number to decimal). Essentially, I first convert the inputted binary number into a decimal number, and then use the derived decimal number to do a conversion into hexadecimal.