Compute the sum of the digits of an integer
This solution works by first receiving the integer in the form of a string so that a string loop may be performed on each character (or each digit). The value of each character is converted into an integer and then added to an arrayList of type Integer. After the string loop is completed, the array list contains as its elements all the digits of the received input integer. A for-loop is then performed on the arrayList whereby each element of the arrayList is added to an int variable called sum (we are adding each digit). This sum variable is then outputted.